improved Sign In test structure in Core package
