+++ packages/core/tests/mfa-totp-sign-in.test.ts
@@ -1,14 +1,14 @@
 import faker from '@faker-js/faker'
+import { afterAll, afterEach, beforeAll, describe, expect, test } from 'vitest'
 import { interpret } from 'xstate'
 import { waitFor } from 'xstate/lib/waitFor'
 import { createAuthMachine } from '../src/machines'
 import { Typegen0 } from '../src/machines/index.typegen'
 import { BASE_URL } from './helpers/config'
 import {
-  signUpConflictErrorHandler,
-  signUpInternalErrorHandler,
-  signUpNetworkErrorHandler,
-  signUpWithSessionHandler
+  mfaTotpInternalErrorHandler,
+  mfaTotpInvalidOtpHandler,
+  mfaTotpNetworkErrorHandler
 } from './helpers/handlers'
 import server from './helpers/server'
 import customStorage from './helpers/storage'
@@ -41,12 +41,12 @@ afterEach(() => {
 })
 
 test(`should fail if network is unavailable`, async () => {
-  server.use(signUpNetworkErrorHandler)
+  server.use(mfaTotpNetworkErrorHandler)
 
   authService.send({
-    type: 'SIGNUP_EMAIL_PASSWORD',
-    email: faker.internet.email(),
-    password: faker.internet.password(15)
+    type: 'SIGNIN_MFA_TOTP',
+    ticket: `mfaTotp:${faker.datatype.uuid()}`,
+    otp: faker.random.numeric(6).toString()
   })
 
   const state: AuthState = await waitFor(authService, (state: AuthState) =>
@@ -55,7 +55,7 @@ test(`should fail if network is unavailable`, async () => {
 
   expect(state.context.errors).toMatchInlineSnapshot(`
       {
-        "registration": {
+        "authentication": {
           "error": "OK",
           "message": "Network Error",
           "status": 200,
@@ -65,12 +65,12 @@ test(`should fail if network is unavailable`, async () => {
 })
 
 test(`should fail if server returns an error`, async () => {
-  server.use(signUpInternalErrorHandler)
+  server.use(mfaTotpInternalErrorHandler)
 
   authService.send({
-    type: 'SIGNUP_EMAIL_PASSWORD',
-    email: faker.internet.email(),
-    password: faker.internet.password(15)
+    type: 'SIGNIN_MFA_TOTP',
+    ticket: `mfaTotp:${faker.datatype.uuid()}`,
+    otp: faker.random.numeric(6).toString()
   })
 
   const state: AuthState = await waitFor(authService, (state: AuthState) =>
@@ -79,7 +79,7 @@ test(`should fail if server returns an error`, async () => {
 
   expect(state.context.errors).toMatchInlineSnapshot(`
       {
-        "registration": {
+        "authentication": {
           "error": "internal-error",
           "message": "Internal error",
           "status": 500,
@@ -88,51 +88,55 @@ test(`should fail if server returns an error`, async () => {
     `)
 })
 
-test(`should fail if either email or password is incorrectly formatted`, async () => {
-  // Scenario 1: Providing an invalid email address with a valid password
+test(`should fail if MFA ticket is not provided or invalid`, async () => {
   authService.send({
-    type: 'SIGNUP_EMAIL_PASSWORD',
-    email: faker.internet.userName(),
-    password: faker.internet.password(15)
+    type: 'SIGNIN_MFA_TOTP',
+    ticket: '',
+    otp: faker.random.numeric(6).toString()
   })
 
-  const emailErrorSignInState: AuthState = await waitFor(
-    authService,
-    (state: AuthState) => !!state.value
+  const noTicketState: AuthState = await waitFor(authService, (state: AuthState) =>
+    state.matches({ authentication: { signedOut: { failed: 'server' } } })
   )
 
-  expect(
-    emailErrorSignInState.matches({
-      authentication: { signedOut: { failed: { validation: 'email' } } }
-    })
-  ).toBeTruthy()
+  expect(noTicketState.context.errors).toMatchInlineSnapshot(`
+      {
+        "authentication": {
+          "error": "no-mfa-ticket",
+          "message": "No MFA ticket has been provided",
+          "status": 10,
+        },
+      }
+    `)
 
-  // Scenario 2: Providing a valid email address with an invalid password
   authService.send({
-    type: 'SIGNUP_EMAIL_PASSWORD',
-    email: faker.internet.email(),
-    password: faker.internet.password(2)
+    type: 'SIGNIN_MFA_TOTP',
+    ticket: faker.datatype.uuid(),
+    otp: faker.random.numeric(6).toString()
   })
 
-  const passwordErrorSignInState: AuthState = await waitFor(
-    authService,
-    (state: AuthState) => !!state.value
+  const invalidTicketState: AuthState = await waitFor(authService, (state: AuthState) =>
+    state.matches({ authentication: { signedOut: { failed: 'server' } } })
   )
 
-  expect(
-    passwordErrorSignInState.matches({
-      authentication: { signedOut: { failed: { validation: 'password' } } }
-    })
-  ).toBeTruthy()
+  expect(invalidTicketState.context.errors).toMatchInlineSnapshot(`
+      {
+        "authentication": {
+          "error": "invalid-mfa-ticket",
+          "message": "MFA ticket is invalid",
+          "status": 10,
+        },
+      }
+    `)
 })
 
-test(`should fail if email has already been taken`, async () => {
-  server.use(signUpConflictErrorHandler)
+test(`should fail if TOTP is invalid`, async () => {
+  server.use(mfaTotpInvalidOtpHandler)
 
   authService.send({
-    type: 'SIGNUP_EMAIL_PASSWORD',
-    email: faker.internet.email(),
-    password: faker.internet.password(15)
+    type: 'SIGNIN_MFA_TOTP',
+    ticket: `mfaTotp:${faker.datatype.uuid()}`,
+    otp: faker.random.numeric(6).toString()
   })
 
   const state: AuthState = await waitFor(authService, (state: AuthState) =>
@@ -141,37 +145,35 @@ test(`should fail if email has already been taken`, async () => {
 
   expect(state.context.errors).toMatchInlineSnapshot(`
       {
-        "registration": {
-          "error": "email-already-in-use",
-          "message": "Email already in use",
-          "status": 409,
+        "authentication": {
+          "error": "invalid-otp",
+          "message": "Invalid or expired OTP",
+          "status": 401,
         },
       }
     `)
 })
 
-test(`should succeed if email and password are correctly formatted`, async () => {
+test(`should succeed if the provided MFA ticket and TOTP were valid`, async () => {
   authService.send({
-    type: 'SIGNUP_EMAIL_PASSWORD',
-    email: faker.internet.email(),
-    password: faker.internet.password(15)
+    type: 'SIGNIN_MFA_TOTP',
+    ticket: `mfaTotp:${faker.datatype.uuid()}`,
+    otp: faker.random.numeric(6).toString()
   })
 
   const state: AuthState = await waitFor(authService, (state: AuthState) =>
-    state.matches({ email: 'awaitingVerification', authentication: { signedOut: 'noErrors' } })
+    state.matches({ authentication: { signedIn: { refreshTimer: { running: 'pending' } } } })
   )
 
-  expect(state.context.user).toBeNull()
-  expect(state.context.errors).toMatchInlineSnapshot('{}')
+  expect(state.context.user).not.toBeNull()
 })
 
-test(`should succeed if email and password are correctly formatted and user is already signed up`, async () => {
-  server.use(signUpWithSessionHandler)
+test(`should succeed if MFA ticket is already in context and TOTP was valid`, async () => {
+  authService.state.context.mfa = { ticket: `mfaTotp:${faker.datatype.uuid()}` }
 
   authService.send({
-    type: 'SIGNUP_EMAIL_PASSWORD',
-    email: faker.internet.email(),
-    password: faker.internet.password(15)
+    type: 'SIGNIN_MFA_TOTP',
+    otp: faker.random.numeric(6).toString()
   })
 
   const state: AuthState = await waitFor(authService, (state: AuthState) =>
